Dynamic commits (#1033)
* build: dynamic commits

* fix: fixing selection algorithm

* fix: compression test

* refactor: moving padArray to common-files

* fix: fixing wallet imports

+++ nightfall-deployer/contracts/Key_Registry.sol
@@ -26,7 +26,7 @@ contract Key_Registry is Ownable, Structures {
       emit VkChanged(_txType);
   }
 
-  function getVerificationKey(TransactionTypes txType) public view returns(uint[] memory) {
-    return vks[txType];
-  }
+    function getVerificationKey(TransactionTypes txType) public view returns (uint256[] memory) {
+        return vks[txType];
+    }
 }
